+++ sdks/java/io/rrio/src/main/java/org/apache/beam/io/requestresponseio/package-info.java
@@ -15,20 +15,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.io.rrio;
 
-import java.io.Serializable;
-
-/**
- * Provided by user and called within {@link org.apache.beam.sdk.transforms.DoFn.Setup} and @{link
- * org.apache.beam.sdk.transforms.DoFn.Teardown} lifecycle methods of {@link Call}'s {@link
- * org.apache.beam.sdk.transforms.DoFn}.
- */
-public interface SetupTeardown extends Serializable {
-
-  /** Called during the {@link org.apache.beam.sdk.transforms.DoFn}'s setup lifecycle method. */
-  void setup() throws UserCodeExecutionException;
-
-  /** Called during the {@link org.apache.beam.sdk.transforms.DoFn}'s teardown lifecycle method. */
-  void teardown() throws UserCodeExecutionException;
-}
+/** Package provides Beam I/O transform support for safely reading from and writing to Web APIs. */
+package org.apache.beam.io.requestresponseio;
