+++ sdks/java/io/rrio/src/test/java/org/apache/beam/io/rrio/SetupTeardownTest.java
@@ -20,6 +20,10 @@ package org.apache.beam.io.rrio;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
 
+import org.apache.beam.io.requestresponseio.SetupTeardown;
+import org.apache.beam.io.requestresponseio.UserCodeExecutionException;
+import org.apache.beam.io.requestresponseio.UserCodeQuotaException;
+import org.apache.beam.io.requestresponseio.UserCodeTimeoutException;
 import org.apache.beam.sdk.coders.VarIntCoder;
 import org.apache.beam.sdk.testing.TestPipeline;
 import org.apache.beam.sdk.transforms.Create;
